x87 optimizations (#62)
Currently we do the slowest thing possible for x87:
- Load `top` from memory
- Load st(n) relative to `top`
- If we modify `top`, write it to memory
- Also writeback st(n)

This results in any x87 operation taking multiple instructions to
accommodate the x87 stack model

---

The idea is to statically allocate the x87 registers in the RISC-V FPRs.
- If there's a push/pop, we need to rotate the allocated registers with
a parallel move
- ft0 <- ft1, ft1 <- ft2, ft2 <- ft3, ..., ft7 <- ft0 (or the other way
around)
    - Also bump TOP in memory
- This way ft0 is always mapped to st0 (ie. the first register relative
to TOP)
- Load FPU state in dispatcher
- If we need to switch to MMX state, keep it inside a block
    - Set using_mmx for signal handlers
    - Move fprs to mmx regs (keep TOP in mind)
    - At end of block, move back to fprs and unset using_mmx
- This way we don't have confusion in the dispatcher which will always
writeback the fpu regs

---

In theory this should cause operations like `fadd st2, st3` to be mapped
to a single RISC-V instruction, while push/pops will have to perform at
least 9 moves, but no load/stores. In theory this should be faster.

---

Other notes:
- There's ways to optimize MMX (different dispatchers based on whether
the block uses MMX or x87?) but that's for another day
- The fp array in ThreadState has the values of our ft0-ft7 registers.
When we save using FXSAVE we probably have to account for the value of
TOP to place each fp value in the correct place
- We can defer pushing/popping within a block but it could mess with
state reconstruction