Vulkan+Wayland thunks (#67)
"Thunks" are special guest libraries that call into guest code. We
overlay these libraries so that when the program tries to open
libvulkan.so for example it will open our thunk libvulkan instead, which
has some magic instructions that we've programmed to call into host
code.

---

This PR adds support for `libvulkan.so` and `libwayland-client.so`. It's
not possible to use Vulkan on wayland without thunking
libwayland-client, because it installs callbacks (that are in x86-64
code) that host Vulkan will call. When thunking libwayland-client, we
capture these callback installations and replace them with a trampoline
that will fixup the arguments and call the guest code (ie. enter the
dispatcher, recompile it, run it, return).

---

- [x] Verify callbacks are working fine
- [x] Add the rest of the Vulkan signatures up to 1.3
- [x] Run vulkaninfo
- [x] Run vkcube
- [x] Run hello-wayland
- [x] Run supertuxkart
- [ ] Check if zink works